+++ CHANGELOG.md
@@ -19,4 +19,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Create project
   - Update project
   - Delete project
-  - Updating channels for a project
+  - Updating channel configuration for a project (undocumented)
+  - Retrieve channel configuration for a project (undocumented)